# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: App API
  version: 0.1.0
  summary: Boilerplate HTTP API — auth, RBAC, users, roles, notifications, realtime
  description: |
    Starter HTTP API for the Next.js + Go boilerplate.

    **In scope:** authentication, global RBAC, platform users/roles, notifications,
    Web Push (VAPID), passkeys, and Centrifugo realtime tokens.

    **Out of scope:** tenants, workspaces, CRM, inbox, commerce, AI, tickets, reports.
    Those product modules were stripped. Live Go routes (`internal/modules/**/routes.go`)
    are the source of truth if a path below is missing from the router.

    **Roles:**
    - `super_admin` — system role; assigned via `make create-super-admin`
    - `organization_user` — system role for business owners/staff
    - Custom roles via `/v1/platform/roles`

    **Auth:** Bearer JWT access tokens + opaque refresh tokens.
    See `docs/auth.md` for the permission matrix and token claims.
    Error catalog: `docs/api/errors.md`.

    **Conventions:**
    - Success/error envelope: `{success,data|error,meta.request_id}` on all `/v1/*`.
    - Lists: `{items,total,limit,offset}` under `data` (`limit`/`offset`/`q`/`sort`; never `page`/`search`).
    - Probes `/healthz` and `/readyz` are **not** enveloped.
    - See `API_CONVENTIONS.md` and `docs/api/errors.md`.
  contact:
    name: App
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: http://127.0.0.1:8080
    description: Local development
...
paths:
...
  /v1/app/config:
    get:
      tags: [Health]
      operationId: getAppConfig
      summary: Public app config
      description: Unauthenticated bootstrap payload (auth methods, branding).
      security: []
      responses:
        "200":
          description: Config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeStatus"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    ErrorDetail:
      type: object
      required: [message]
      properties:
        field: { type: string }
        message: { type: string }
        code: { type: string }
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
          examples:
            - VALIDATION_ERROR
            - INVALID_RESET_CODE
            - INVALID_VERIFICATION_CODE
            - INVALID_CREDENTIALS
            - UNAUTHENTICATED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - PASSWORD_RESET_FAILED
            - INTERNAL_ERROR
        message:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/ErrorDetail"
    ResponseMeta:
      type: object
      required: [request_id]
      properties:
        request_id: { type: string }
    Error:
      type: object
      required: [success, error, meta]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          $ref: "#/components/schemas/ErrorBody"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    StatusPayload:
      type: object
      required: [status]
      properties:
        status: { type: string }
...
    EnvelopeStatus:
      type: object
      required: [success, data, meta]
      properties:
        success: { type: boolean, enum: [true] }
        data: { $ref: "#/components/schemas/StatusPayload" }
        meta: { $ref: "#/components/schemas/ResponseMeta" }
...